+++ systems/evolution.js
@@ -11,7 +11,6 @@ import { GAME_INFO } from '../config/gameConfig.js';
 import { t } from '../lang.js';
 import { addXP } from './player.js';
 import { applyOrganEffects, applyHiddenOrganEffects, showOrganSelection, showHiddenOrganSelection, getOrganSlotsUsed } from './organs.js';
-import { showFloatingText } from './combat.js';
 import { applyMutationEffects, saveMutationData, getMutationUpgradeCost, upgradeMutation, initMutationSkills, _syncMutationSkillPoints, _saveMutationSkills } from './mutation.js';
 import { showTooltip, hideTooltip, buildEndGameOverlay } from './ui.js';
 import { pausePlayTimer, resumePlayTimer } from './gameFlow.js';